+++ lib/get_arch_core.dart
@@ -7,8 +7,9 @@ export 'domain/i_entity.dart';
 export 'domain/env_config.dart';
 export 'domain/extension.dart';
 export 'domain/live_data.dart';
-export 'package:dartz/dartz.dart' show Either, Unit;
+export 'package:dartz/dartz.dart' show Either, Unit, right, left;
 export 'package:rxdart/rxdart.dart';
+export 'package:meta/meta.dart';
 
 export 'interface/i_dto.dart';